refactoring

+++ d_server/inc/gpio.h
@@ -1,5 +1,5 @@
 void get_sensor_state(int lamp[], int ac[], int sp[], int so[]);
-void set_ac_state(short value, short opt);
-void set_lamp_state(short value, short opt);
+void set_ac_state(int value, int opt);
+void set_lamp_state(int value, int opt);
 void configure_pins();
